Fixed Generics for lists and added proper testing
