--- src/JSONParser/FieldSerializer/FieldDeserializer.java
@@ -1,9 +1,0 @@
-package JSONParser.FieldSerializer;
-
-import JSONParser.JSONValues.JSONElement;
-import JSONParser.JSONValues.JSONObject;
-
-@FunctionalInterface
-public interface FieldDeserializer <T>{
-    T deserialize(JSONElement object);
-}
