+++ src/main/JSONParser/FieldSerializer/FieldDeserializer.java
@@ -1,0 +1,9 @@
+package main.JSONParser.FieldSerializer;
+
+import main.JSONParser.JSONValues.JSONElement;
+import main.JSONParser.JSONValues.JSONObject;
+
+@FunctionalInterface
+public interface FieldDeserializer <T>{
+    T deserialize(JSONElement object);
+}
