+++ src/main/JSONParser/JSONValues/JSONArray.java
@@ -1,4 +1,4 @@
-package JSONParser.JSONValues;
+package main.JSONParser.JSONValues;
 
 import java.util.*;
