+++ src/main/JSONParser/JSONValues/JSONObject.java
@@ -1,4 +1,4 @@
-package JSONParser.JSONValues;
+package main.JSONParser.JSONValues;
 
 import java.util.*;
 import java.util.Map.*;
