+++ src/main/JSONParser/JSONValues/JSONString.java
@@ -1,34 +1,33 @@
-package JSONParser.JSONValues;
+package main.JSONParser.JSONValues;
 
 import java.util.Objects;
 
-public class JSONValue implements JSONElement {
+public class JSONString implements JSONElement {
     private final String val;
 
-    public JSONValue(Object val) {
-        this.val = String.valueOf(val);
+    public JSONString(String val) {
+        this.val = val;
     }
 
     public String serialize() {
-        return val;
+        return String.format("\"" + val + "\"");
     }
 
     @Override
     public Object getValue() {
         return val;
     }
-
     @Override
     public String toString() {
         return serialize();
     }
 
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        JSONValue that = (JSONValue) o;
-
+        JSONString that = (JSONString) o;
         return val.equals(that.val);
     }
