+++ src/main/JSONParser/JSONValues/JSONValue.java
@@ -1,33 +1,34 @@
-package JSONParser.JSONValues;
+package main.JSONParser.JSONValues;
 
 import java.util.Objects;
 
-public class JSONString implements JSONElement {
+public class JSONValue implements JSONElement {
     private final String val;
 
-    public JSONString(String val) {
-        this.val = val;
+    public JSONValue(Object val) {
+        this.val = String.valueOf(val);
     }
 
     public String serialize() {
-        return String.format("\"" + val + "\"");
+        return val;
     }
 
     @Override
     public Object getValue() {
         return val;
     }
+
     @Override
     public String toString() {
         return serialize();
     }
 
-
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        JSONString that = (JSONString) o;
+        JSONValue that = (JSONValue) o;
+
         return val.equals(that.val);
     }
