+++ src/main/ParserCombinators/Serializer/Serializer.java
@@ -1,6 +1,6 @@
-package ParserCombinators.Serializer;
+package main.ParserCombinators.Serializer;
 
-import JSONParser.JSONValues.JSONObject;
+import main.JSONParser.JSONValues.JSONObject;
 
 public interface Serializer<T> {
     void deserialize(JSONObject object, T o);
