+++ src/main/Test/Class1.java
@@ -1,11 +1,11 @@
-package Test;
+package main.Test;
 import java.util.ArrayList;
 import java.util.List;
 
 
 public class Class1<T> {
     public Class2 x;
-    public List<List<Integer>> b;
+    public List<List<List<Integer>>> b;
 //    public T val;
     public Class3<T,Integer> c3;
     public Class1() {
